requirements.txt
branca==0.5.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dnspython==1.16.0
Flask==2.0.1
folium==0.12.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
pymongo==3.12.0
python-dotenv==0.20.0
requests==2.28.1
typing-extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.1
zipp==3.8.1
